Use binutils 2.24, and Ubuntu-12.04 for Windows builds
To be closer to what we currently have in gitian, use binutils 2.24.

We also change tho OS to Ubuntu 12.04 for Windows builds, except for the
tor-browser bundle step, where we are staying with 14.10. We are using
14.10 in the tor-browser step because the version of python-pefile we
use fails on Ubuntu 12.04. The reason for using a recent version of
python-pefile is that the older versions don't work on x86_64. And we
are using an x86_64 container because docker does not support i386
containers. The problem is not present in gitian where we are using an
i386 VM. An other possible solution could be to build a recent version
of python.