Bug 23425: fix build when there is no terminal
Set process.terminal to false when building, but keep it to true when we
open a debugging shell.

https://github.com/opencontainers/runtime-spec/blob/master/config.md#process

To do that we update the runc config.json in runc/remote_exec rather
than only once in runc/remote_start.

We also redirect runc stderr to /dev/null to avoid some messages saying
"stdin: is not a tty" during the build.