Bug 23454: restore stderr in build logs
In #23425 we redirected the output from runc start to /dev/null to
avoid annoying messages saying "stdin: is not a tty". However this also
removes stderr from build logs.

The command producing the `stdin: is not a tty` message is the `chown -R`
command from runc/remote_put, so we now only hide stderr from that
part while keeping it for the runc command running the build script.