Bug 23039: fix detection of runc version
In the previous patch, we wrongly assumed that all runc 0.1.1 versions
exited with 0 when given an unknown command, and used that to detect the
version. It seems it is not true on all systems.

We are now using the 'runc --version' command to find the version of
runc.

We don't support versions older than 0.1.1.