Bug 23585: fix build with runc version 1.0.0~rc2
Some distributions are packaging runc version 1.0.0~rc2, which seems to
be half between runc 0.1.1 and runc 1.0.0.

This version requires the same command line parameters as version 1.0.0,
however it requires a config.json in the same format as 0.1.1.

The output from `runc --version` on 1.0.0~rc2 is:
runc version spec: 1.0.0-rc2-dev

So we add a var/runc_spec100 function which is true when the runc
version spec is exactly 1.0.0 (as returned by runc stable 1.0.x
releases), and use it in projects/common/runc-config.json.