Bug 24341: update platform.arch to avoid error with some runc versions
The runc versions after this commit:
  https://github.com/opencontainers/runc/commit/1cd050244e004bc9940787f888b4570328703100
and before this commit:
  https://github.com/opencontainers/runc/commit/e1146182a8cebb5a6133a9e298a5e4acf99652e9
cannot run containers where platform.arch is set to anything else than
amd64 when running on amd64.

The platform.arch is not used for anything so we can set it to amd64 in
all cases to avoid the error with that runc version.