Reset timestamp when no_build_id is selected
The defaut timestamp value will use the commit time of the selected
commit for the project, which will require cloning the git repository if
it is not present. When we use the no_build_id target to display a script,
we usually don't care about such details, so we set timestamp to 0 to
avoid unnecessary cloning.