Bug 26010: Change Snowflake rendezvous to use the Azure domain front
https://bugs.torproject.org/22782#comment:7

This also uses the newer standalone broker on the backend (rather than
the one based on App Engine), but that should be invisible to users.

Patch by David Fifield

I converted the mac file to a unix file format as well using `dos2unix`
to get rid of the ^M characters that started to show up.