Bug 26057: make it easy to see commit used in nightly builds
In order to make it easy to see which commit was used, we include the
abbreviated commit hash in the filename of all components that are based
on master in the nightly build. To do that, we set git_hash to master,
and version (which is included in the filename) to '[% c("abbrev") %]".