Bug 26039: <profiledir>/preferences/extension-overrides.js will not be loaded in ESR 60
The infrastructure for loading extension-overrides.js no longer
exists in ESR60, so these prefs had to be moved.

Unfortunately, we can't just dump the contents of extension-overrides.js
into 000-tor-browser.js.  The settings are actually partially built in
tor-browser-build to conditionally include various bridge strings as well
as localization settings.  So, we piggy back off this existing build
process and just append the contents of the partially generated
extension-overrids.js to 000-tor-browser.js.