Bug 26319: Don't package up the whole Tor Browser in the `mach package` step
This commit sets `mach build stage-package` instead of `mach package` in
the Firefox build; this helps us to use the useful properties of `mach
package` but avoids creating the final DMG or EXE during the Firefox
build which we don't want as we do that later when building Tor Browser.

We no longer need to patch `installer.py` (bug 26205) but we still need
the patch for the uninstaller otherwise the Firefox build fails trying
to find the required NSIS files.

The no-dmg.patch for macOS is also not required and is removed (bug 24632).