Bug 26475: Disable building Rust with Thin LTO
Building Rust with Thin LTO enabled leads to Tor Browser builds on macOS
and probably Linux not being reproducible. The exact reason for that is
unknown at the moment, although it seems fixed testing nightly Rust
source tarballs as of mid-September 2018.

We therefore disable Thin LTO for now by setting `codegen-units` to `1`.