Bug 27827: pass -tmpdir to the linker in snowflake build
The update to go 1.11.1 fixed part of #27827, namely the
/tmp/go-buildRANDOMNUM paths embedded in the binary. However there is
still a single /tmp/go-link-RANDOMNUM/go.o path remaining -- it comes
from here:
https://github.com/golang/go/blob/go1.11.1/src/cmd/link/internal/ld/lib.go#L935-L946
The random path affects the Go build ID even if we overwrite the random
digits in the binary.

Luckily there's a -tmpdir flag that lets us override the random path.