Bug 26843: Get l10n-changesets.json from tor-browser.git
The json file used by projects/firefox-locale-bundle is only available
from https://product-details.mozilla.org/ after a build has been
shipped, which is too late for us. However, it is also available from
the firefox sources tree, so we can get it from our tor-browser.git
clone instead.

The format of this json file is a little different, so we also need to
adapt the get_hg_hash script.