Bug 26148: Update binutils to 2.31.1
With older versions of binutils, the order in which .o files were
included in an .a archive did not affect the output when linking a
program or library, which is no longer the case with recent versions of
binutils. To avoid reproducibility issues we patch the libtool version
included in gcc to make the creation `libstdc++.a` reproducible.