Bug 29981: Add option to build without using containers
We do that by adding support for a no_containers target which defines
the new option var/container/global_disable in rbm.conf. We already had
the option var/container/use_container doing something similar to this
new option, however it can not be overriden in rbm.conf as we define it
in each project's configuration, which has priority over rbm.conf.