Bug 30039: Use the target_prepend option in https-everywhere
Update rbm for #30039, adding support for the target_prepend and
target_append options in input_files.

In projects/https-everywhere/config we now use the target_prepend
option instead of the target option for the python input file.
This does not change the output of the build. However this makes it
possible in the future to check the channel or platform we build for
in the python project (which was not possible before as we were using
the target option, replacing all the initial targets).