Bug 30549: Add script to remove expired and revoked sub-keys from a keyring file
The tools/keyring/drop-expired-sub-keys script can be used to drop all
expired and revoked sub-keys from a keyring.

We also add the script tools/keyring/list-all-keyrings which can be used
to list all the keys included in all the keyring files, to make it
easier to review if any key needs to be removed.